Added CowBytes for deserializing borrowed or owned
Same as using ArcBytes but it avoids the arc. This allows this library
to be used in all situations without neeing to plop in serde_bytes for
deserializing borrowed bytes without introducing the overhead of the
Arc.

This allows users to pick ArcBytes if they are deserializing a structure
they plan on cloning multiple times, or CowBytes if they are
deserializing for one instance.