Fixed edge case with Read when in EOF state
In an attempt to free up memory as eagerly as possible, the Read
implementation would notice once ArcBytes was EOF and replace the
buffer with None.

Future read operations, however, would yield interesting results, as the
position and end fields now pointed beyond the slice. `into_owned()` in
this situation would yield a panic.

The proper solution was to clear start/end as well. I've also refactored
this to behave the same in both Read/read_bytes, one of the issues I had
when trying to reproduce the bug.